% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clipp}
\alias{clipp}
\title{clip function.}
\usage{
clipp(x)
}
\arguments{
\item{x}{A vector or matrix.}
}
\value{
A vector or matrix same as the input.
}
\description{
Limit the number not to be too large or too small.
}
