% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Fprep.AIWKME}
\alias{Fprep.AIWKME}
\title{Cox proportional hazards model for eta-free terms in the (S)AIWKME estimator.}
\usage{
Fprep.AIWKME(datalist, t0)
}
\arguments{
\item{datalist}{A list used to calculate the (S)AIWKME estimator including treatment named \code{a}, observed time named \code{obs.t}, censoring indicator (0, censored) named \code{delta}, and baseline covariates used to assign treatment named \code{l}.}

\item{t0}{A predetermined t.}
}
\value{
A list including the augmented terms in the numerator with treatment all to 1 named \code{gamma.num.1} and all to 0 named \code{gamma.num.0} and in the denominator with treatment all to 1 named \code{gamma.den.1} and all to 0 named \code{gamma.den.0}; \code{gamma.num.1} and the others are matrix with ordered observed time as rows and patients as columns. More details can be found in references.
}
\description{
Cox proportional hazards model for eta-free terms in the (S)AIWKME estimator.
}
\details{
More details can be found in references, \code{\link[otrKM]{AIWKME}}, and \code{\link[otrKM]{Genetic.optim.AIWKME}}.
}
\references{
{Jiang, R., Lu, W., Song, R., and Davidian, M. (2017) On estimation of optimal treatment regimes for maximizing t‐year survival probability. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, \bold{79:} 1165-1185. DOI:10.1111/rssb.12201}
}
