% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Fps.IVE}
\alias{Fps.IVE}
\title{Logistic regression for observed treatment and instrument used for the (S)IVE estimator.}
\usage{
Fps.IVE(datalist)
}
\arguments{
\item{datalist}{A list used to calculate the (S)IVE estimator including treatment named \code{a}, observed time named \code{obs.t}, censoring indicator (0, censored) named \code{delta}, and baseline covariates used to assign treatment named \code{l}.}
}
\value{
A list including the probability of receiving instrument given baseline covariates named \code{fzl} and the difference between fal1 and fal0 named \code{deltal}, where fal0 denotes the probability of receiving treatment given baseline covariates and instrument equaling 0, and fal1 denotes the probability of receiving treatment given baseline covariates and instrument equaling 1.
}
\description{
Logistic regression for observed treatment and instrument used for the (S)IVE estimator.
}
\details{
More details can be found in references, \code{\link[otrKM]{IVE}}, and \code{\link[otrKM]{Genetic.optim.IVE}}.
}
\references{
{Xia, J., Zhan, Z., Zhang, J. (2022) An anti-confounding method for estimating optimal regime in a survival context using instrumental variable. Under Review.}
}
