% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IWKME.R
\name{IWKME}
\alias{IWKME}
\title{The (S)IWKME estimator.}
\usage{
IWKME(eta, datalist, ps, t0, smooth = TRUE)
}
\arguments{
\item{eta}{The parameters of the regime.}

\item{datalist}{A list used to calculate the (S)IWKME estimator including treatment named \code{a}, observed time named \code{obs.t}, censoring indicator (0, censored) named \code{delta}, and baseline covariates used to assign treatment named \code{l}. \code{\link[otrKM]{Fps.IWKME}} can produce \code{ps} by positing logistic model. Notice that all the data in the datalist should be ordered by observed time.}

\item{ps}{A list including the probability of receiving treatment given baseline covariates named \code{fal}.}

\item{t0}{A predetermined time.}

\item{smooth}{A logic variable indicating wether a smoothed estimator should be used.}
}
\value{
Estimated potential survival probability given eta and t0.
}
\description{
Given a predetermined t0 and eta, calculate t0-year potential survival probability based on the (S)IWKME estimator.
}
\details{
More details can be found in references.
}
\examples{
# load data
data(simulation)
simulation=simulation[order(simulation$Survival),]

# convert the data into a datalist
datalist=list(z=simulation$Instrument,a=simulation$Treatment,
              obs.t=simulation$Survival,delta=simulation$Status,
              l=cbind(simulation$Covariate1,simulation$Covariate2))

# calculate ps
ps=Fps.IWKME(datalist)

# predetermined t0 and eta
t0=5
eta=c(1,2,3)

IWKME(eta, datalist, ps, t0, smooth=TRUE)
}
\references{
{Jiang, R., Lu, W., Song, R., and Davidian, M. (2017) On estimation of optimal treatment regimes for maximizing t‐year survival probability. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, \bold{79:} 1165-1185. DOI:10.1111/rssb.12201}
}
