% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulation}
\format{
An object of class \code{data.frame} with 500 rows and 6 columns.
}
\usage{
simulation
}
\description{
A simulation data to help understand and implement the functions in the package. 'Instrument' denotes the binary instrumental variable. 'Treatment' denotes the binary treatment. 'Survival' denotes the observed survival time. 'Status' denotes whether the the data is censoring where '0' means censoring. 'Covariate' denotes the baseline characteristics.
}
\examples{
  head(simulation)
}
\keyword{datasets}
